Fixes the color of the buttons for LSA
On the analysis dashboard, the styles of the buttons for the LSA template were wrong. To fix it, the $theme variable has been used in c-input.scss but it implied to have the management and admin settings files to have the variable set.

In addition, the JS needed to know which template is loaded so it can add the correct classes to the buttons. As a consequence, in the front, a template variable has been added to window with either "lsa" or "fa" as value