Fixes an issue with the list of sites
The list of sites wouldn't load if the user would be seeing, for example, the list of users and would use the quick links component to go back to the list of sites. The issue is that a rails redirection is not done and that the dispatcher wouldn't load the router for this route.