Improves the quality of the scaled down images
If we directy scale down the images with a canvas, the implementation of the browsers returns a low quality image. One possible solution is to scale down by steps but it makes the code even more complex. Instead, by using a library, we can skip all of this and have a good quality image.

The images are also "uploaded" 1.5x larger that what we actually need to they look nicer on Retina desktops.