Feature/wysiwyg images improvement (#143)
* Improves the quality of the scaled down images

If we directy scale down the images with a canvas, the implementation of the browsers returns a low quality image. One possible solution is to scale down by steps but it makes the code even more complex. Instead, by using a library, we can skip all of this and have a good quality image.

The images are also "uploaded" 1.5x larger that what we actually need to they look nicer on Retina desktops.

* Fixes the position of the caption for narrow images

The caption is now aligned with the image, to its right, instead of aligned to the right of the container.

* Lets the user changes the size of the images/widgets

A new button has been added in the toolbar to toggle between 3 size: the width of the content, the previous default width of the images/widgets and the width of the cover images.

By default, the images/widgets are displayed with the "normal" size (i.e. smaller than before).