Fixes an issue w/ the precision of the numeric filters
The precision would be at least of 1, which would lead to too many options to choose from when the difference between the minimum and maximum value would be big. It would eventually lead the browser to a crash.

With this commit, the precision is computed so always 100 values to choose from are displayed to the user.