Improvements to the widget creation/edition (#293)
* Open the advanced mode with the vega config of the editor

On the widget creation page, when the user switches to the advanced mode, the textarea contains the `vegaConfig` object of the editor.

* Move the advanced mode toggle closer to the widget-editor

The new component `ToggleSwitcher` has been created for the UI.

* Let the user update a "standard" widget to an advanced one

* Save the theme in the advanced mode

The theme defined in the wysiwyg takes precedence over the theme selector. If no theme is defined in the wysiwyg, the theme of the selector will be picked.